+++ VUI/Config/Layouts/_Castbars.lua
@@ -50,6 +50,20 @@ function Layout:OnEnable()
                     tooltip = 'Display cast time on castbar',
                     column = 4
                 },
+                latency = {
+                    key = 'latency',
+                    type = 'checkbox',
+                    label = 'Latency',
+                    tooltip = 'Display latency information on the left side of the castbar',
+                    column = 4
+                },
+                targetname = {
+                    key = 'targetname',
+                    type = 'checkbox',
+                    label = 'Target Name',
+                    tooltip = 'Display target name on the right side of the castbar',
+                    column = 4
+                },
                 targetCastbar = {
                     key = 'targetCastbar',
                     type = 'checkbox',